Add files via upload
Kathleen here are sections 1 and 5. There were some sections that I just didn't feel like I knew how to answer and felt that we should collaborate or I can look into further, but wanted your feedback. I hope this looks OK as I kind of felt like I was BSing quite a bit and a lot of the stuff didn't really pertain to this short project.